requirements.txt
pyaml>=16.12.2
nltk>=3.2.2
